Reorganizing devices into constants and profiles
