+++ sensorkit/__init__.py
@@ -1,8 +1,18 @@
-__all__ = [ 'config', 'controls', 'datastructures', 'devices', 'devicetree', 'meters' ]
+__all__ = [ 'config',
+           'constants',
+           'controls',
+           'datastructures',
+           'devices',
+           'devicetree',
+           'meters',
+           'profiles'
+]
 
 from . import config
+from . import constants
 from . import controls
 from . import datastructures
 from . import devices
 from . import devicetree
 from . import meters
+from . import profiles
